requirements.txt
passlib
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
flask-oidc==1.4.0
Flask-SQLAlchemy==2.3.2
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==4.1.3
okta==0.0.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-slugify==1.2.5
requests==2.19.1
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.12
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.14.1
